Former-commit-id: dd37083099596311a6546b884837ab988c3b6e8f

+++ applications/programs/classify_kerdensom/classify_kerdensom_main.cpp
@@ -39,8 +39,8 @@ public:
     double         eps;          // Stopping criteria
     unsigned       iter;         // Iteration number
     bool           norm;         // Normalize?
-    int            xdim;         // X-dimension (-->)
-    int            ydim;         // Y-dimension
+    size_t         xdim;         // X-dimension (-->)
+    size_t         ydim;         // Y-dimension
     double         reg0;         // Initial reg
     double         reg1;         // Final reg
     std::string    layout;       // layout (Topology)
@@ -247,7 +247,7 @@ public:
         MetaData vectorHeaderIn, vectorHeaderOut, vectorContentOut;
         vectorHeaderIn.read(formatString("vectorHeader@%s",fn_in.c_str()));
         vectorHeaderOut.setColumnFormat(false);
-        int size, vectorSize;
+        size_t size, vectorSize;
         size_t idIn=vectorHeaderIn.firstObject();
         size_t idOut=vectorHeaderOut.addObject();
         vectorHeaderIn.getValue(MDL_XSIZE,size,idIn);
@@ -256,7 +256,7 @@ public:
         vectorHeaderOut.setValue(MDL_YSIZE,size,idOut);
         vectorHeaderIn.getValue(MDL_ZSIZE,size,idIn);
         vectorHeaderOut.setValue(MDL_ZSIZE,size,idOut);
-        vectorHeaderOut.setValue(MDL_COUNT,(size_t)myMap->size(),idOut);
+        vectorHeaderOut.setValue(MDL_COUNT,myMap->size(),idOut);
         vectorHeaderIn.getValue(MDL_CLASSIFICATION_DATA_SIZE,vectorSize,idIn);
         vectorHeaderOut.setValue(MDL_CLASSIFICATION_DATA_SIZE,vectorSize,idOut);
         vectorHeaderOut.write(formatString("vectorHeader@%s_vectors.xmd",fn_root.c_str()),MD_APPEND);
